Added bind_property_to_widget_value
This is starting to simplify the updating of the interface.  I used a strategy similar to Apple's Cocoa pre-swift.  I use the TkVariables and I connect the value of one widget (checkbox, entry, etc...) and bind them to the property of another widget.  By setting the property with @properties, we can manage the change of the value on the widget, and update accordingly in the host object.