Many different changes regarding window geometry
If Base.debug = True, then you will get a message explaining the geometry situation when you place a widget.  I really needed this to understand the details of why a widget was resizing or why the window was not resizing.

App used to set the geometry to a default size (1000x500).  This was a mistake because it prevented the window from adjusting its size when propagate() was 1 and compeltely short-circuited the Tkinter layout manager.  Now it is None if it is not set.

Also, after() should really be called from the class instead of widget to deal with a deleted widget before the call is completed.