FileViewer is improving fast
Many improvements in quick succession to the FileViewer.
It will now allow a custom column that can be used for calculations for instance.

We can limit the depth of the listdir (done manually, easy).
For now, it is too difficult to update the records when clicked to dig into the tree, but it is coming: I figured out a way to update the items without deleting the whole tree, but this requires more work than expected.