Update canvas_app.py
Reworked significantly around a "bug" in ImagingPath: when calculating the field stop, we need to stop after the image position! If there is an aperture after, then it is not relevant.
So we need to do a two-pass strategy: 1) find image position without apertures, then get path up to that position with aperture.